gui: move game settings into a submenu - game menu
