+++ include/lang.h
@@ -269,6 +269,9 @@ enum _STR_IDS {
     _STR_CACHE_HDD_GAME_LIST,
     _STR_ENABLE_NOTIFICATIONS,
     _STR_NOTIFICATIONS,
+    _STR_GAME_MENU,
+    _STR_GAME_SETTINGS_SAVED,
+    _STR_GAME_SETTINGS_REMOVED,
 
     LANG_STR_COUNT
 };
