+++ include/menusys.h
@@ -81,6 +81,7 @@ typedef struct menu_list
 void menuInit();
 void menuEnd();
 void menuReinitMainMenu(void);
+void menuInitGameMenu(void);
 
 void menuAppendItem(menu_item_t *item);
 
@@ -98,9 +99,11 @@ void menuSaveConfig();
 void menuRenderMain();
 void menuRenderMenu();
 void menuRenderInfo();
+void menuRenderGameMenu();
 void menuHandleInputMain();
 void menuHandleInputMenu();
 void menuHandleInputInfo();
+void menuHandleInputGameMenu();
 
 // Sets the selected item if it is found in the menu list
 void menuSetSelectedItem(menu_item_t *item);
